refactoring ThingIFAPI
