+++ thingif/src/main/java/com/kii/thingif/ThingIFAPI.java
@@ -31,11 +31,11 @@ import com.kii.thingif.core.exception.ThingIFException;
 import com.kii.thingif.core.exception.ThingIFRestException;
 import com.kii.thingif.core.exception.StoredThingIFAPIInstanceNotFoundException;
 import com.kii.thingif.core.exception.UnsupportedActionException;
-import com.kii.thingif.core.exception.UnsupportedSchemaException;
+import com.kii.thingif.exception.UnsupportedSchemaException;
 import com.kii.thingif.core.gateway.EndNode;
 import com.kii.thingif.core.gateway.Gateway;
 import com.kii.thingif.core.gateway.PendingEndNode;
-import com.kii.thingif.core.internal.GsonRepository;
+import com.kii.thingif.internal.GsonRepository;
 import com.kii.thingif.core.internal.http.IoTRestClient;
 import com.kii.thingif.core.internal.http.IoTRestRequest;
 import com.kii.thingif.schema.Schema;
