+++ thingif/src/main/java/com/kii/thingif/exception/UnsupportedSchemaException.java
@@ -1,4 +1,6 @@
-package com.kii.thingif.core.exception;
+package com.kii.thingif.exception;
+
+import com.kii.thingif.core.exception.ThingIFException;
 
 /**
  * Thrown when an unsupported schema is detected.
